+++ lib/database/sync_db.dart
@@ -35,7 +35,9 @@ class SyncDatabase extends _$SyncDatabase {
           ),
         );
 
-  final bool inMemoryDatabase;
+  SyncDatabase.connect(super.connection) : super.connect();
+
+  bool inMemoryDatabase = false;
 
   Future<int> updateOutboxItem(OutboxCompanion item) {
     return (update(outbox)..where((t) => t.id.equals(item.id.value)))
@@ -94,4 +96,8 @@ class SyncDatabase extends _$SyncDatabase {
 
   @override
   int get schemaVersion => 1;
+}
+
+SyncDatabase getSyncDatabase() {
+  return SyncDatabase.connect(getDatabaseConnection(syncDbFileName));
 }