+++ lib/sync/encryption_messages.dart
@@ -5,14 +5,6 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 part 'encryption_messages.freezed.dart';
 
 @freezed
-class EncryptStringMessage with _$EncryptStringMessage {
-  factory EncryptStringMessage({
-    required String plaintext,
-    required String b64Secret,
-  }) = _EncryptStringMessage;
-}
-
-@freezed
 class DecryptStringMessage with _$DecryptStringMessage {
   factory DecryptStringMessage({
     required String encrypted,
@@ -20,14 +12,6 @@ class DecryptStringMessage with _$DecryptStringMessage {
   }) = _DecryptStringMessage;
 }
 
-@freezed
-class EncryptFileMessage with _$EncryptFileMessage {
-  factory EncryptFileMessage({
-    required String b64Secret,
-    required File inputFile,
-    required File encryptedFile,
-  }) = _EncryptFileMessage;
-}
 
 @freezed
 class DecryptFileMessage with _$DecryptFileMessage {