+++ test/encryption_string_test.dart
@@ -6,7 +6,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:lotti/sync/encryption.dart';
 
 void main() {
-  test('Check AES GCM encryption/decryption roundtrip', () async {
+  test('Check AES GCM encryption/decryption round trip', () async {
     final originalFile = File('test_resources/test.txt');
     final testString = await originalFile.readAsString();
     final b64Secret = Key.fromSecureRandom(32).base64;